requirements.txt
flask
pyspark
kafka-python
pymongo
cassandra-driver
pandas
scikit-learn
